improve code portability

+++ bld/wipfc/cpp/btmpblk.hpp
@@ -43,7 +43,7 @@ public:
     void write( std::FILE* out ) const;
     STD1::uint32_t compress( std::FILE* in );
     unsigned int totalSize() const
-    { return sizeof( STD1::uint16_t ) + ( data.size() + 1 ) * sizeof( STD1::uint8_t ); };
+    { return static_cast< unsigned int >( sizeof( STD1::uint16_t ) + ( data.size() + 1 ) * sizeof( STD1::uint8_t ) ); };
 private:
     enum lzwBits {
         INITBITS = 9,
